[MINOR] Fix static code analysis warnings, part I

+++ src/main/java/org/apache/sysds/runtime/compress/colgroup/dictionary/IdentityDictionary.java
@@ -529,7 +529,7 @@ public class IdentityDictionary extends ADictionary {
 
 	@Override
 	public double getSparsity() {
-		return 1.0d / (double) nRowCol;
+		return 1d / nRowCol;
 	}
 
 	@Override
