+++ src/main/java/org/apache/sysds/runtime/instructions/fed/CastFEDInstruction.java
@@ -131,7 +131,7 @@ public class CastFEDInstruction extends UnaryFEDInstruction {
 
 		// Construct output local.
 		FrameObject out = ec.getFrameObject(output);
-		out.getDataCharacteristics().set(mo1.getNumRows(), mo1.getNumColumns(), (int) mo1.getBlocksize(), mo1.getNnz());
+		out.getDataCharacteristics().set(mo1.getNumRows(), mo1.getNumColumns(), mo1.getBlocksize(), mo1.getNnz());
 		FederationMap outMap = mo1.getFedMapping().copyWithNewID(fr2.getID());
 		List<Pair<FederatedRange, FederatedData>> newMap = new ArrayList<>();
 		for(Map.Entry<FederatedRange, FederatedData> pair : outMap.getMap()) {
